# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Cat:
    properties:
      breed:
        example: Siamese
        type: string
      id:
        type: integer
      name:
        example: Tom
        type: string
      salary:
        example: 1000
        type: integer
      years_of_experience:
        example: 5
        type: integer
    required:
    - breed
    - name
    type: object
  domain.CatRequest:
    properties:
      breed:
        example: Siamese
        type: string
      name:
        example: Tom
        type: string
      salary:
        example: 1000
        type: integer
      years_of_experience:
        example: 5
        type: integer
    required:
    - breed
    - name
    type: object
  domain.LoginRequest:
    properties:
      email:
        example: email@example.com
        type: string
      password:
        example: Password12345!
        maxLength: 50
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  domain.Mission:
    properties:
      cat_id:
        example: 1
        type: integer
      completed:
        example: false
        type: boolean
      id:
        type: integer
      notes:
        example: Lorem ipsum
        type: string
      targets:
        items:
          $ref: '#/definitions/domain.Target'
        type: array
    required:
    - cat_id
    - targets
    type: object
  domain.MissionRequest:
    properties:
      cat_id:
        example: 1
        type: integer
      completed:
        example: false
        type: boolean
      notes:
        example: Lorem ipsum
        type: string
      targets:
        items:
          $ref: '#/definitions/domain.Target'
        type: array
    required:
    - cat_id
    - targets
    type: object
  domain.Response:
    properties:
      message:
        example: response message
        type: string
    type: object
  domain.Target:
    properties:
      completed:
        example: false
        type: boolean
      country:
        example: USA
        type: string
      id:
        type: integer
      mission_id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      notes:
        example: Lorem ipsum
        type: string
    required:
    - country
    - mission_id
    - name
    type: object
  domain.UserRequest:
    properties:
      email:
        example: email@example.com
        type: string
      password:
        example: Password12345!
        maxLength: 50
        minLength: 8
        type: string
      username:
        example: username
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
host: localhost:8888
info:
  contact:
    email: raitermark@proton.me
    name: Mark Raiter
  description: Docs for SpyCat API
  title: SpyCat API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs user in
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Login
      tags:
      - Auth
  /auth/logout:
    post:
      description: Logs user out
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register user
      parameters:
      - description: User data
        in: body
        name: Register_request
        required: true
        schema:
          $ref: '#/definitions/domain.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User ID
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Register user
      tags:
      - Auth
  /cats:
    get:
      consumes:
      - application/json
      description: Get all cats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Cat'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all cats
      tags:
      - Cat
    post:
      consumes:
      - application/json
      description: Create cat
      parameters:
      - description: Cat data
        in: body
        name: Create_cat_request
        required: true
        schema:
          $ref: '#/definitions/domain.CatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Cat ID
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Create cat
      tags:
      - Cat
  /cats/{id}:
    delete:
      consumes:
      - application/json
      description: Delete cat by ID
      parameters:
      - description: Cat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete cat by ID
      tags:
      - Cat
    get:
      consumes:
      - application/json
      description: Get cat by ID
      parameters:
      - description: Cat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Cat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Get cat by ID
      tags:
      - Cat
    put:
      consumes:
      - application/json
      description: Update cat by ID
      parameters:
      - description: Cat ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cat data
        in: body
        name: Update_cat_request
        required: true
        schema:
          $ref: '#/definitions/domain.CatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Update cat by ID
      tags:
      - Cat
  /missions:
    get:
      consumes:
      - application/json
      description: Get missions
      produces:
      - application/json
      responses:
        "200":
          description: Missions
          schema:
            items:
              $ref: '#/definitions/domain.Mission'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Get missions
      tags:
      - Mission
    post:
      consumes:
      - application/json
      description: Create mission
      parameters:
      - description: Mission data
        in: body
        name: Create_mission_request
        required: true
        schema:
          $ref: '#/definitions/domain.MissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Mission ID
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "406":
          description: Not Acceptable
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Create mission
      tags:
      - Mission
  /missions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete mission
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete mission
      tags:
      - Mission
    get:
      consumes:
      - application/json
      description: Get mission by ID
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Mission
          schema:
            $ref: '#/definitions/domain.Mission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Get mission by ID
      tags:
      - Mission
    patch:
      consumes:
      - application/json
      description: Complete mission
      parameters:
      - description: Mission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Complete mission
      tags:
      - Mission
  /missions/{mission_id}/cats/{cat_id}:
    patch:
      consumes:
      - application/json
      description: Assign mission to cat
      parameters:
      - description: Cat ID
        in: path
        name: cat_id
        required: true
        type: integer
      - description: Mission ID
        in: path
        name: mission_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Assign mission to cat
      tags:
      - Mission
  /missions/{mission_id}/targets/{target_id}:
    patch:
      consumes:
      - application/json
      description: Add target to mission
      parameters:
      - description: Mission ID
        in: path
        name: mission_id
        required: true
        type: integer
      - description: Target ID
        in: path
        name: target_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Add target to mission
      tags:
      - Target
  /targets/{id}:
    patch:
      consumes:
      - application/json
      description: Complete target
      parameters:
      - description: Target ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Complete target
      tags:
      - Target
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"